\docType{package}
\name{evcpick}
\alias{evcpick}
\alias{evcpick-package}
\title{Package \code{evcpick}}
\description{
  The package provides functions for rank-based versions of
  three estimators for the Pickands dependence function of
  a bivariate extreme value copula: the Pickands estimator,
  the Caperaa-Fougeres-Genest estimator, and the
  Buecher-Dette-Volgushev estimator.
}
\references{
  C. Genest and J. Segers (2009), The Annals of Statistics.
  A. Buecher, H. Dette, and S. Volgushev (2011), The Annals
  of Statistics.
}
\seealso{
  \code{\link{evcPickands}}, \code{\link{evcCFG}},
  \code{\link{evcBDV}}
}

